refactoring
rename the module for dp utilities

update dpnp/dpctl checkers

reusing daal4py sklearn_check_version
